+++ packages/components/tests/unit/helpers/validateCssAndMatchSnapshot.ts
@@ -12,12 +12,17 @@ export const validateCssAndMatchSnapshot = (css: string) => {
 
   expect(css).not.toMatch('. {'); // Invalid css which was produced before
 
+  console.log('componentNameMETA__', {
+    componentName,
+    componentMeta,
+  });
   validateVisibilityStyle(cssObject);
   validateSlottedStyles(cssObject, componentTagName);
   validateHoverMediaQuery(cssObject);
   validatePreventFoucOfNestedElementsStyle(
     cssObject,
-    componentMeta && Array.isArray(componentMeta.nestedComponents) && componentMeta.nestedComponents.length > 0
+    (componentMeta && Array.isArray(componentMeta.nestedComponents) && componentMeta.nestedComponents.length > 0) ||
+      ['input-base'].includes(componentName)
   );
 
   // Validations for components only